refactor directory structure and make setup.py
