+++ src/__init__.py
@@ -1,0 +1,3 @@
+import pytest
+import math
+import json
