+++ tests/test_LogicalExpressionClass.py
