+++ tests/test_TableClass.py
